readMe update

+++ Wallet-App/Model/History.swift
@@ -10,7 +10,7 @@ import Foundation
 struct Expenses {
     let recieverName: String
     let date: Date
-    let amount: Double
+    let amount: Float
     let sector: EconomicSector
 }
